# travelea-api-spec_1.0.0.yaml
openapi: 3.0.0
info:
  description: SENG302 Team 400 - Travel EA
  version: "1.0.0-oas3"
  title: Travel EA
  contact:
    email: s302team400@cosc.canterbury.ac.nz
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  '/trip/privacy':
    put:
      tags:
      - Trips
      summary: Updates an existing trip's privacy
      security:
      - UserToken: []
      parameters:
      - in: header
        name: JWT-Auth
        required: true
        description: Authentication cookie.
        schema:
          type: string
      requestBody:
        $ref: '#/components/requestBodies/putTrip'
      responses:
        '200':
          description: Successfully updated trip
        '400':
          description: Invalid data supplied, ErrorResponse returned detailing specific validation failures
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  '/destination':
    post:
      tags:
      - Destinations
      summary: Add a new destination
      description: |
        Inserts a destination into the database, that is owned by the current user
      security:
      - UserToken: []
      parameters:
      - in: header
        name: JWT-Auth
        required: true
        description: Authentication cookie.
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Destination'
      responses:
        '200':
          description: Successfully added destination
        '400':
          description: Invalid data supplied, ErrorResponse returned detailing specific validation failures
        '401':
          description: Unauthorized
...
servers:
- url: 'http://csse-s302g4.canterbury.ac.nz:8080/api'
components:
  requestBodies:
...
    putTrip:
      content:
        application/json:
          schema:
            type: object
            required:
            - userId
            - tripId
            - tripDataCollection
            properties:
              userId:
                type: integer
                minimum: 1
                format: int64
                description: References a the id of an existing user
              tripId:
                type: integer
                minimum: 1
                format: int64
                description: References a the id of the trip that is being updated
              tripDataCollection:
                type: array
                items:
                  $ref: '#/components/schemas/TripData'
      description: 'Trip data necessary for new trip to be added. Note: destination field only needs to include id property, nothing else.'
  schemas:
    User:
      type: object
      required:
      - username
      - password
      properties:
        id:
          type: integer
          minimum: 1
          format: int64
          description: >-
            Automatically generated when user is signed up, should never be
            specified manually
        username:
          type: string
          format: email
          example: dave@gmail.com
        password:
          type: string
          minLength: 6
          example: MeowMeow123
          description: >-
            Password is only plaintext when uploading user object for the first
            time, or updating the user password
        salt:
          type: string
          description: Randomly generated salt used for hashing password
        admin:
          type: boolean
          description: Indicates whether or not this user is an admin
        creationDate:
          type: string
          format: ISO Datetime
          description: Generated automatically at the time of creation of the user
...
    CountryDefinition:
      type: object
      required:
      - id
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
          example: France
...
    Destination:
      type: object
      required:
      - name
      - type
      - district
      - latitude
      - longitude
      - country
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          description: Generated automatically when the destination is first uploaded
        user:
          $ref: '#/components/schemas/User'
          description: The user that owns this destination. If marked as private, this will be whichever user uploaded the destination, otherwise if marked as public (and has been used by another user) then this will be the master admin.
        isPublic:
          type: boolean
          example: false
          description: Whether or not this destination is considered a private destination of the user that added it or whether it is publicly accessible
        name:
          type: string
          example: Eiffel Tower
        type:
          type: string
          example: Monument
          description: 'Type of destination, e.g monument, national park, etc.'
        district:
          type: string
          example: Paris
        latitude:
          type: number
          minimum: -90
          maximum: 90
          example: 48.858093
        longitude:
          type: number
          minimum: -180
          maximum: 180
          example: 2.294694
        country:
          $ref: '#/components/schemas/CountryDefinition'
...
    TripData:
      type: object
      required:
      - position
      - destination
      description: >-
        IMPORTANT NOTE: When specifying the destination this trip data
        corresponds to, only the destination's id field must be filled in.
      properties:
        guid:
          type: integer
          format: int64
          minimum: 1
          description: Generated automatically when the tripdata is first uploaded
        position:
          type: integer
          format: int64
          minimum: 0
          description: >-
            The position during the trip at which this tripdata occurs. I.e
            first destination visited has position 0, then position 1, and so
            on.
        destination:
          $ref: '#/components/schemas/Destination'
        arrivalTime:
          type: string
          format: ISO date-time
          example: '2019-04-01T23:26:00.000'
        departureTime:
          type: string
          format: ISO date-time
          example: '2019-04-20T10:30:00.000'
...